app/requirements.txt
Flask==3.0.0
python-dotenv
google-cloud-datastore==2.15.1